+++ src/components/real/communication/antenna_radiation_pattern.cpp
