+++ src/components/real/communication/antenna_radiation_pattern.hpp
@@ -3,8 +3,8 @@
  * @brief Class to manage antenna radiation pattern
  */
 
-#ifndef S2E_COMPONENTS_COMMUNICATION_ANTENNA_RADIATION_PATTERN_HPP_
-#define S2E_COMPONENTS_COMMUNICATION_ANTENNA_RADIATION_PATTERN_HPP_
+#ifndef S2E_COMPONENTS_REAL_COMMUNICATION_ANTENNA_RADIATION_PATTERN_HPP_
+#define S2E_COMPONENTS_REAL_COMMUNICATION_ANTENNA_RADIATION_PATTERN_HPP_
 
 #include <library/math/constants.hpp>
 #include <string>
@@ -62,4 +62,4 @@ class AntennaRadiationPattern {
   std::vector<std::vector<double>> gain_dBi_;  //!< Antenna gain table [dBi]
 };
 
-#endif  // S2E_COMPONENTS_COMMUNICATION_ANTENNA_RADIATION_PATTERN_HPP_
+#endif  // S2E_COMPONENTS_REAL_COMMUNICATION_ANTENNA_RADIATION_PATTERN_HPP_
