+++ src/components/real/communication/initialize_antenna.hpp
@@ -3,10 +3,10 @@
  * @brief Initialize function for Antenna
  */
 
-#ifndef S2E_COMPONENTS_COMMUNICATION_INITIALIZE_ANTENNA_HPP_
-#define S2E_COMPONENTS_COMMUNICATION_INITIALIZE_ANTENNA_HPP_
+#ifndef S2E_COMPONENTS_REAL_COMMUNICATION_INITIALIZE_ANTENNA_HPP_
+#define S2E_COMPONENTS_REAL_COMMUNICATION_INITIALIZE_ANTENNA_HPP_
 
-#include <components/communication/antenna.hpp>
+#include <components/real/communication/antenna.hpp>
 
 /*
  * @fn InitAntenna
@@ -16,4 +16,4 @@
  */
 Antenna InitAntenna(const int antenna_id, const std::string file_name);
 
-#endif  // S2E_COMPONENTS_COMMUNICATION_INITIALIZE_ANTENNA_HPP_
+#endif  // S2E_COMPONENTS_REAL_COMMUNICATION_INITIALIZE_ANTENNA_HPP_
