+++ src/components/real/communication/initialize_ground_station_calculator.cpp
