+++ src/simulation/spacecraft/sample_spacecraft/sample_components.hpp
@@ -14,7 +14,7 @@
 #include <components/real/aocs/initialize_star_sensor.hpp>
 #include <components/real/aocs/initialize_sun_sensor.hpp>
 #include <components/real/cdh/on_board_computer.hpp>
-#include <components/communication/initialize_antenna.hpp>
+#include <components/real/communication/initialize_antenna.hpp>
 #include <components/examples/example_change_structure.hpp>
 #include <components/examples/example_i2c_controller_for_hils.hpp>
 #include <components/examples/example_i2c_target_for_hils.hpp>
